third_party/WebKit/LayoutTests/http/tests/performance-timing/paint-timing/first-contentful-image.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test</title>
</head>
<body>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script>
    async_test(function (t) {
        var observer = new PerformanceObserver(
            t.step_func(function (entryList) {
                var entries = entryList.getEntries();
                assert_equals(entries.length, 2,
                    "There should be two paint timing instances.");
                assert_equals(entries[0].entryType, "paint",
                    "Expected entryType to be: paint.");
                assert_equals(entries[0].duration, 0,
                    "Expected duration to be: 0.");

                assert_equals(entries[1].entryType, "paint",
                    "Expected entryType to be: paint.");
                assert_equals(entries[1].duration, 0,
                    "Expected duration to be: 0.");
                observer.disconnect();
                t.done();
            })
        );
        observer.observe({entryTypes: ["paint"]});

    }, "First contentful paint fires due to image render.");
    </script>

    <script>
    window.onload = function() {
        var img = document.createElement("IMG");
        img.src = "resources/circles.png";
        document.getElementById('image').appendChild(img);
    }
    </script>

    <div id="image"></div>
</body>
</html>
